updated repositories to use jpa instead of crud changed id type from int to Long. added create faculty endpoint

+++ src/main/java/com/lms/lmsapi/repository/StudentCategoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.lms.lmsapi.entity.StudentCategory;
 
 @Repository
-public interface StudentCategoryRepository extends JpaRepository<StudentCategory, Integer>{
+public interface StudentCategoryRepository extends JpaRepository<StudentCategory, Long>{
     
 }
