+++ src/main/java/com/lms/lmsapi/service/FacultyService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface FacultyService {
     
     List<Faculty> getFaculties();
-    Faculty getFaculty(int Id);
+    Faculty getFaculty(Long Id);
     Faculty createFaculty(Faculty faculty);
     Faculty updateFaculty(Faculty faculty);
-    void deleteFaculty(int FacultyId);
+    void deleteFaculty(Long FacultyId);
 }
